django7182nd2n/requirements.txt
django==2.0
django-threadlocals
pymysql
click
requests
xlrd==1.2.0
Flask==1.1.2
django-cors-headers==2.4.0
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.3
scikit-learn==1.0.2
SQLAlchemy==1.3.20
joblib==1.3.2
seaborn==0.12.2

hdfs==2.7.0
paramiko==3.4.0
pyspark==3.3.2
pyhive==0.7.0
thrift-sasl==0.4.3
mysql-connector-python==8.0.33
django-pyodbc==1.1.3
pycryptodome
